Fixes based on feedback.
No longer check if we're on a terminal every single time a progress update is to
be printed. Check occurs once and is passed to the function.

Add re-export of HideProgress option to mirror re-export of HideSuccesses.

Only set the cursor column when we're displaying progress output.

Store progress update intervals in a separate IORef.

Change obscure function name to something more obvious.

Fix limit value to correct number, remove let binding and use in place.